pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
 name = "clasper-control-plane-starter"
 version = "0.1.0"
 description = "Minimal Control Plane starter with OpenClaw-style plugins."
 requires-python = ">=3.10"
 dependencies = [
   "fastapi>=0.110.0",
   "uvicorn>=0.27.0",
   "pyjwt>=2.8.0",
   "pydantic>=2.6.0",
 ]

 [project.optional-dependencies]
 dev = [
   "pytest>=8.0.0",
   "httpx>=0.27.0",
 ]

 [project.entry-points."clasper.plugins"]
 # Example: "sample_plugin = plugins.sample_plugin.plugin:register"

[tool.setuptools.packages.find]
where = ["src"]

 [tool.pytest.ini_options]
 testpaths = ["tests"]

